+++ NEWS.md
@@ -6,6 +6,9 @@ Version 1.0.3b0 - dev
 ---------------------
 
 - Release notes: https://ezdxf.mozman.at/release-v1-0.html
+- NEW: [#833](https://github.com/mozman/ezdxf/issues/833)
+  logging non-unique entity handles when loading a DXF document as warnings, auditing 
+  the document may fix this issue
 
 Version 1.0.2 - 2023-02-15
 --------------------------